requirements.txt
contourpy==1.3.3
CTkListbox==1.5
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
